+++ crates/client/db/src/storage_updates.rs
@@ -3,7 +3,11 @@ use std::collections::HashMap;
 use mp_convert::field_element::FromFieldElement;
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce, PatriciaKey};
 use starknet_api::hash::StarkFelt;
-use starknet_core::types::{DeclaredClassItem, DeployedContractItem, NonceUpdate, ReplacedClassItem, StateUpdate};
+use starknet_api::state::StorageKey;
+use starknet_core::types::{
+    ContractStorageDiffItem, DeclaredClassItem, DeployedContractItem, NonceUpdate, ReplacedClassItem, StateUpdate,
+    StorageEntry,
+};
 use storage_handler::primitives::contract_class::{
     ClassUpdateWrapper, ContractClassData, ContractClassWrapper, StorageContractClassData,
 };
@@ -12,7 +16,7 @@ use crate::storage_handler::{self, DeoxysStorageError, StorageView, StorageViewM
 
 pub async fn store_state_update(block_number: u64, state_update: StateUpdate) -> Result<(), DeoxysStorageError> {
     let state_diff = state_update.state_diff.clone();
-    let nonce_map: HashMap<ContractAddress, Nonce> = state_update
+    let mut nonce_map: HashMap<ContractAddress, Nonce> = state_update
         .state_diff
         .nonces
         .into_iter()
@@ -26,7 +30,7 @@ pub async fn store_state_update(block_number: u64, state_update: StateUpdate) ->
 
     log::debug!("💾 update state: block_number: {}", block_number);
 
-    let (result1, result2, result3, result4) = tokio::join!(
+    let (result1, result2, result3) = tokio::join!(
         // Contract address to class hash and nonce update
         async move {
             let handler_contract_data = storage_handler::contract_data_mut();
@@ -47,10 +51,14 @@ pub async fn store_state_update(block_number: u64, state_update: StateUpdate) ->
                 let previous_nonce = handler_contract_data.get(&contract_address).unwrap().map(|data| data.nonce);
                 let nonce = match previous_nonce.unwrap_or_default().get().copied() {
                     Some(nonce) => Some(nonce),
-                    None => nonce_map.get(&contract_address).copied(),
+                    None => nonce_map.remove(&contract_address),
                 };
 
                 handler_contract_data.insert(contract_address, (class_hash, nonce)).unwrap()
+            });
+
+            nonce_map.into_iter().for_each(|(contract_address, nonce)| {
+                handler_contract_data.insert(contract_address, (None, Some(nonce))).unwrap()
             });
 
             handler_contract_data.commit(block_number)
@@ -77,15 +85,12 @@ pub async fn store_state_update(block_number: u64, state_update: StateUpdate) ->
         },
         // Block number to state diff update
         async move { storage_handler::block_state_diff().insert(block_number, state_diff) },
-        // Contract address to contract storage update
-        async move { storage_handler::contract_storage_mut().commit(block_number) }
     );
 
-    match (result1, result2, result3, result4) {
-        (Err(err), _, _, _) => Err(err),
-        (_, Err(err), _, _) => Err(err),
-        (_, _, Err(err), _) => Err(err),
-        (_, _, _, Err(err)) => Err(err),
+    match (result1, result2, result3) {
+        (Err(err), _, _) => Err(err),
+        (_, Err(err), _) => Err(err),
+        (_, _, Err(err)) => Err(err),
         _ => Ok(()),
     }
 }
@@ -105,4 +110,24 @@ pub async fn store_class_update(block_number: u64, class_update: ClassUpdateWrap
     );
 
     handler_contract_class_data_mut.commit(block_number)
+}
+
+pub async fn store_key_update(
+    block_number: u64,
+    storage_diffs: &Vec<ContractStorageDiffItem>,
+) -> Result<(), DeoxysStorageError> {
+    let handler_storage = storage_handler::contract_storage_mut();
+
+    for ContractStorageDiffItem { address, storage_entries } in storage_diffs {
+        let contract_address = ContractAddress::from_field_element(address);
+        for StorageEntry { key, value } in storage_entries {
+            let key = StorageKey::from_field_element(key);
+            let value = StarkFelt::from_field_element(value);
+            handler_storage.insert((contract_address, key), value)?;
+        }
+    }
+
+    handler_storage.commit(block_number)?;
+
+    Ok(())
 }